feat: New `privacy` module
To allow for keeping personal information, so not "secrets", but
"sensitive" information out of the public repo. Ofc they will still
exist on the store, but that isint a concern.

This module has to be enabled for the host, then number of checks will
be done that throw errors if something went wrong.