fix(privacy): Fixed ReadOnly attribute
Turns out if you were to disable the privacy modules you'd get and error
due to `my.privacy.data` not being defined resulting OR assigments of values
to fail and throw errors.

This fixes the problem so that its always defined, either with
the actuall data or simply just `{}` when disabled.