refactor(privacy): Defined ssh bootstraping issue
Turns out if i want Nix to use a special key i need to tell it about it,
but before i can tel it about it it tries fetching from a private repo.
So a mechanism and process for doing this initial setting up to avoid
error when building a host first time.