reafctor(privacy): Changed relationships
the old my.privacy.schema is meant to be able to be used and edited as
needed, so if someone wants to add change values with 0 care about
privacy, then it doenst make sense to have it under privacy.

instead made a new `my.hostData` where the data will live and the
`my.privacy` module will when enabled just populate the `my.hostData`
attributes as it did previously with `my.privacy.schema`

can now begin structuring profiles using this interface, making minimal
assumptions of things that need to defined

like "my.hostData.users.main" needs to be defined ect.

which one can either define on a private repo and import using privacy
logic or define it in configuration.nix to host, whatever  they see fit