refactor!: hostData been merged with host
Besides the naming change, the my.host.users now will serve as the
source of truth for what "user" defined users will be created on the
host. For now it simply by default intilises all users that are defined
udner my.host.users and by default they will fail the build, this is
intentional to prevent users accidentally being defined.