refactor(git): made the template system more dynamic
Instead of defining the strucutre and expectations for a template in
user git module, a syste mhas been implemented so the template provides,
what variables i relies on, what assertions it needs set and
dependencies.

This way templates are not required to follow a "global" assertion
table. And also make their own custom assert to ensure the expectations
they have are met