refactor(privacy)!: Changed how privacy is imported
Now instead of doing a very convoluted aproach of evaluating the privacy
file for host manually, its imported as a standalone module, which
makes the process more clean.

This also allows for setting options in the privacy module, altough i
think this defeats the purpose of the public part of the repo, but its
there and available to be used.