feat(git): Implemented proper git config template
Expanded the earlier made template, and made it work. Templates are
responsible for specifying packages the packages the user needs and also
what assertions to have checked.

I simply just created a config ill be using but for different profiles
if ever needed, special ones could be made and be set up per user on a
host.