feat(nvim): each user can now have their own nvim config
current implementation assumes all different profiles are defined in the
same repositoy. If it has to be this way im a little unsure,
theoretically implementation could be adjusted so users define their own
inputs and it tries parsing the flake that way. But again im going to
keep it simple and close to my own use case for now.