refactor(nvim): added dev mode to nvim
When wanting to configure nvim, and see the results rightaway, a new
apprach implemented. Define environment variable NIX_NVIM_DEV=1 and the
path specified for the user will be used

Also the git url alias was causing problems with lazy nvim pulling from
repo and needing ssh key. so removed it.