feat: implemented dynamic fs configuration
Now hosts define how its hardware looks like and defines a link to given
a mount point the disk type its on can be deduced. This is a draft
implementation that essentially only is implemented for my use case and
has not been defined for other setups than mine. But it easly can be
expanded.

Also added swapfile