feat: hyprland & configuration implementation
Added hyprland (UWSM) as the desktop environment.

Along side a lot of pre configured applications and settings. The idea
here is that the configuration part is rather easily swappable with a
different one as long as the right information in the config source is
provided.

Its a minimal clone of what home manager does.

Might have bugs or be limited in functionaly, but as everything else
will be worked on and improved over time