feat: added new host, Hermes
This is a more modern laptop i have and im happy to say that it worked
installing this configuration on it.

I still need to refine the installetion process are there are multiple
reboots needed to setup secure-boot, swapfile/hibernation, then also
gpg cauing problems since it expects values which are not accessible
when ssh is not setup for privacy.