reafactor: restrucutred repo
The previous structure i followed i realized was not simple to navigate
and was "spread" in implementation, especially with
my.host.users.<user> define options and their actual implementation. So
i changed that firstly default.nix now only serves to import modules.
And options are defined in the file they also are implemented. This is
what ill be following moving forward